hardpwn/glitchlab/pigpio_glitch_transport.py
"""
Pi-based simple glitcher using a GPIO to toggle a MOSFET or power switch.
This is a hardware-dependent routine. You must wire the MOSFET gate to the chosen pin with proper level-shifting.
"""
import time
try:
    import RPi.GPIO as GPIO
except Exception:
    GPIO = None

class PiGpioGlitchTransport:
    def __init__(self, db=None, power_pin=18):
        self.db = db
        self.pin = power_pin
        if GPIO:
            GPIO.setmode(GPIO.BCM)
            GPIO.setup(self.pin, GPIO.OUT)
            GPIO.output(self.pin, GPIO.LOW)

    def glitch_voltage(self, pulse_ns, delay_ns):
        if GPIO is None:
            return {'status':'no_gpio'}
        time.sleep(delay_ns/1e9)
        # pulse MOSFET gate high for duration
        GPIO.output(self.pin, GPIO.HIGH)
        time.sleep(pulse_ns/1e9)
        GPIO.output(self.pin, GPIO.LOW)
        return {'status':'ok'}

    def glitch_clock(self, pulse_ns, delay_ns):
        # Implement clock injection toggling an injected pin; placeholder
        return {'status':'ok_placeholder'}

    def glitch_reset(self, pulse_ns, delay_ns):
        # Implement reset pin pulse; placeholder
        return {'status':'ok_placeholder'}
