requirements.txt
spidev
smbus
pyserial
RPi.GPIO
sqlite-utils
